Add support for auto restart on change
Using [nodemon](https://github.com/remy/nodemon) we can have the app auto restart by itself when a file changes.

Obviously our app is a tiny demo thing so this is not needed. But this project in support of using GitHub Actions with [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) and helping me understand them better. The API also uses nodemon and references it in its Dockerfile. So, to help keep things similar we add it to this project.