Add initial version of a CI workflow (#1)
This includes a step to help us understand how events are interrupted by GitHub. Specifically, pushing to `main` and `pushing` to a branch for a PR. We'll only want a Docker image to build when

- `main` is updated
- a new `tag` is added

We don't want Docker Images built for every push on a branch that isn't `main`.