Add integration with SonarCloud
This is what we use on the SROC Charging Module API and we want to investigate how this step works in conjunction with whatever we plan to do with Docker.

We're explororing the other [GitHub action workflow events](https://docs.github.com/en/actions/reference/events-that-trigger-workflows) so if we make a change to how and when things get kicked off we want to know SonarCloud will still work.