Split CI workflow into separate jobs (#3)
Based on our reading it looks like we could create a Docker workflow that would be triggered by the [GitHub Action `workflow_run` event](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#workflow_run).

This means we can start to break up the actions and trigger them at specific points. Another benefit of doing this is it also looks like the [recommended way](https://securitylab.github.com/research/github-actions-preventing-pwn-requests/) to run a build when changes come in but still access repo secrets.

Our Dependabot PR's started failing some time ago because they could no longer access the `SONAR_TOKEN` in a repo's secrets.

Note - We add an untested workflow action that relies on the event `workflow_run` with this PR because as per the docs

> Note: This event will only trigger a workflow run if the workflow file is on the default branch.

So, we need to get the action merged before it will trigger. Hopefully, we can then edit and test with it in a future PR.