Update docker workflow
We're scrapping the idea of trying to trigger the docker workflow after the CI one as it's over complicating things.

Instead we've taken what GitHub generates via the UI when you say you want to build and push a Docker image to a registry with some minor tweaks. We'll use that as our starter-for-ten whilst we work out how everything works.