Add build args to docker workflow
When we build our SROC Charging Module API image we depend on 2 build args being passed in. These are then 'baked' into the image so that the API can return the git sha and docker tag of the version of the API running. We use this info for diagnostic purposes.

This change to our demo project we hope will replicate setting and using these build args when the image is built by GitHub actions.