Working final version of build args solution
This version we have build args working as we need and being passed as we expect to the `docker build` call. We're happy to just use the info generated by [metadata-action](https://github.com/docker/metadata-action) as that will ensure we remain consistent with the labels it will assign to our image.