Add push to DockerHub to workflow
Having applied what we have learnt here to [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) we then hit a blocker. For the [Defra Organisation](https://github.com/DEFRA) the default settings for packages is that they are private and that access needs to be explicit.

We're looking at container registry usage as an org but until 'there can be only one' there is a good chance we might need to support pushing to multiple registries.

So, this change updates the Docker workflow to push to both registries.