Use better logic to set dockerhub tag(s)
Using sed we realised we could do a better version of setting the DockerHub tag. But doing it this way if you push a 'tag' then we realised DockerHub would only get one of them (a git tag results in `v*.*.*` plus `latest` docker tags).

If we ditched sed and just used the [metadata-action](https://github.com/marketplace/actions/docker-metadata-action) for a second time we would get a result that was acceptable to [tag-push-action](https://github.com/marketplace/actions/tag-push-action#using-with-dockermetadata-action) and that would include the same docker tags.

Or so we thought! Though the docs for tag-push-action say it supports multiple destinations there is currently an [outstanding issue](https://github.com/akhilerm/tag-push-action/issues/88) which prevents this. So, in the interim we have implemented a workaround that means we only push an image at a time to DockerHub, but we check if the 'latest' needs doing when the event is a new tag being added.