Improved version of DockerHub integration
The previous update to `docker.yml` had the same outcome as this one. Whether the change was a push to `main` or a new git tag all Docker tags generated by [Docker Metadata action](https://github.com/marketplace/actions/docker-metadata-action) would be matched when we push to DockerHub.

But we then spied that [Docker Build and Push action](https://github.com/marketplace/actions/build-and-push-docker-images) had a document that covered [pushing to multi-registries](https://github.com/docker/build-push-action/blob/master/docs/advanced/push-multi-registries.md). Essentially, having logged into all the registries that would be pushed to as long as the tags you passed it contained alll the registries, it would handle things.

If we could get this to work

- no more working round an issue
- less action dependencies
- skip an unneccessary step - previous version required pulling then pushing (it was based on copying between registries)
- simpler workflow

This change is the result of our efforts. We are doing everything we wanted but with all the benefits outline above.